server/requirements.txt
Django==2.0.7
psycopg2-binary==2.7.5
django-debug-toolbar==1.9.1
gunicorn==19.7.1
djangorestframework==3.8.2
fcm_django==0.2.21
djangorestframework-jwt
Pillow==5.1.0